Wait longer for Instruments to launch when retrying
Subliminal script originally waits 30 seconds when launching
Instruments and the app. If nothing has been written after these 30
seconds, the Simulator is killed and it tries it all again — if the
retry count limit has not yet been reached.

In older machines, 30 seconds may not be enough for Instruments and the
app to launch. Let’s implement a sort of exponential fall back and
increase the timeout duration after each try: 30 seconds, 45 seconds, 1
minutes and then 2 minutes.