Guard against nil case for accessibilityElementAtIndex:
If the accessibility protocol methods are not properly implemented (as
in custom controls for example), we might have a case where
accessibilityElementAtIndex: will return nil. This causes a crash in
slChildAccessibilityElementsFavoringSubviews: with a nil insertion
into an array. The fix is to guard against adding nil element to the
children array.