Subliminal aborts the run if the simulator is in an inconsistent state. (closes #147)
In certain environments like Travis, `instruments` intermittently hangs.
When this occurs, it seems that the simulator is also in an inconsistent state
such that it can't be rotated, web pages don't load, etc. If we detect that
we are running in such an environment, Subliminal aborts the run after logging
an error. `subliminal-test` detects this error and retries.