Use `instruments`' `-w` flag to enforce the simulator SDK & device type. (resolves #143)
This capability, added in Xcode 5.1, obviates the need to enable/disable
SDKs to force `instruments` to use a particular SDK for testing
(http://openradar.appspot.com/radar?id=3107401) and the need to
specify the `TARGETED_DEVICE_FAMILY` argument to force `instruments`
to use a particular device family for testing
(http://openradar.appspot.com/13607967).

The first components of the configurations respected by the `-w` flag
(the device types) are the same as the items listed in the iOS Simulator's
"Hardware -> Device" menu, so `subliminal-test` remains
backwards-compatible, despite its documentation changing.

Some references to iOS 5.1 have been updated in preparation for our
eventual dropping of 5.1 support (though we still do support it currently).