Allocate pixels with all zero components for bitmap context
We should use calloc here instead of malloc to ensure that the
pixel buffer we use for testing visibility is all clear when
nothing has been drawn yet.  This can be an issue in cases where
renderViewRecursively doesn't actually draw anything.  If nothing
is drawn then this function should return 0, but if we're unlucky
with the random values in the malloced pixel buffer we might see
something "drawn" in the pixel buffer anyway and return non-zero.