Install `appledoc` from binary on Travis.
appledoc 2.1 uses GC and so cannot be compiled using Xcode 5.1,
so we download a pre-built version from the `gh-pages` branch
and manually install it.

The version was pre-built following the instructions [here][1] using
Xcode 5.0.2 and the `appledoc21` formula. I thought that maybe we
could install the resulting bottle [directly][2] but for some reason that
fails.

[1] https://github.com/Homebrew/homebrew/wiki/Bottles
[2] https://github.com/Homebrew/homebrew/wiki/Bottles#local-bottle-usage