Fix the index html when building the docs for release.
There are two problems with directly injecting the README into the autogenerated
main index html page:

1. Appledoc can't handle fenced code blocks and some links.
2. Appledoc puts the class hierarchy all the way below the README.

Now, when we build in release, the Rake task will process the index html itself:

1. Using GitHub's Markdown API to render the README, and injecting the resulting
    html itself;
2. Injecting CSS to override Appledoc's stylesheet and float the class hierarchy to
    the left of the README
    (h/t https://github.com/CocoaPods/cocoadocs.org/blob/master/views/appledoc_stylesheet.scss)

This process requires the Nokogiri gem but the requirement will be invoked
only when necessary, i.e. when the docs are built for Release. We only attempt
to fix the index html up when building for Release to speed Debug builds,
to avoid requiring this gem to be installed in general, and because the appearance
of the index probably doesn't matter as much to developers.