Ensure that the status bar test has completely scrolled to top before finishing.
The app can crash due to a bad-access exception in `UIScrollView`
if a scroll view is deallocated before it finishes scrolling to top in response
to a scroll bar tap.

It's also simpler to test that the scroll view content offset changes as we expect
than that labels in the scroll view do/do not become visible.

Unfortunately, it appears as though Travis requires an explicit wait beyond
the wait-for loop.