`-[UIView slAccessibilityIsVisible]` now reports an element's visibility on-screen, regardless of window. (fixes #212)
That is, `-slAccessibilityIsVisible` will now properly report the visibility of views
in any window, even non-key windows; and `-slAccessibilityIsVisible` will now
properly report whether a view is visible on-screen, not merely within its window
(i.e. it will report `NO` if a view is occluded by the contents of a window above its window).

This removes the need for `SLPickerView` to use `UIAElement.isVisible`
(in order to match pickers used as input views).