Update `simulator-set` for Xcode 6.
In Xcode 6, the Simulator's device type and version are not stored separately
in the Simulator's preferences, but as a unified configuration (a "device")
identified by a UDID. To change the device type and version, we must thus determine
which UDID to set, which we do by examining the mapping provided by
`instruments -s devices`.

Similarly, the Simulator's content and settings are no longer stored
under version-specific directories, but under device-specific directories.
However, it's not safe to delete these directories directly--some devices
will not recreate them on launch. Fortunately, Apple has added a command-line
utility for managing devices.