Log an explicit warning about there not being tests to run.
Rather than just a message.

Also update the copy, because the old message was weird (I'm not sure why it
would say that no tests were focused if we clearly were running with focus);
because it's nicer to distinguish between the user simply not having defined
any tests, vs. them having tried to run tests that can't be run; and to point
them at the documentation to remedy their issue.

Tests have been added for these various cases.