Focused tests exclude other tests even if they cannot be run. (fixes #234)
To make it easier to reason about what focus really means--trying to run
one tests, at the exclusion of all other tests.

No one should have relied on the previous behavior--focus being ignored
if the focused test could not be run--insofar as focusing is temporary,
and focusing an unsupported test would never have caused that test to run.